refactor(react-scheduling): split the write out of the booking form
`onBook` and `onBooked` were two props for one outcome on one component, and
passing both silently ignored the second: "not called when `onBook` took it over"
was the tell. The exclusivity a doc comment had to explain is now in the types.

`AppointmentProposalForm` is every field, the search and the chosen time. It
builds the proposal and hands it to a required `onBook`, writing and announcing
nothing. `AppointmentBookingForm` wraps it and supplies that `onBook`: it posts
`Appointment/$book`, announces the written Appointment and each reserved Slot,
then reports through `onBooked`. Neither offers the other's callback, so mounting
the form that does not write means naming a callback that says the write is
yours. Both are exported, and the one called `AppointmentBookingForm` is still
the one a host mounts with a single prop.

Two behaviours a host taking the write over used to miss come with the move. The
form marks itself booked whenever `onBook` resolves, not only when it wrote, so
taking the write over no longer leaves it offering to book the same time twice.
And a rejected `onBook` is shown as the booking's refusal with every answer kept,
whichever side attempted the write. A failure after the write — `onBooked`
throwing — is still reported where a host's own failures are, since the
appointment exists by then.

The tests split the same way: the proposal form's drive `onBook` directly with no
`$book` stub, and the wrapper's cover the `$book` call, the announcements,
`onBooked` and a refusal. The helpers that drive the form to a bookable state
move to `test-utils/bookingForm` rather than being copied into both files.
